backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-dotenv
pydantic
passlib[bcrypt]
python-jose
python-multipart
argon2_cffi
bcrypt==4.0.1
